alert => callout; fixes #104
