+++ docs/_includes/sidebar.njk
@@ -43,9 +43,6 @@
 <h2>Components</h2>
 <ul>
   <li>
-    <a href="/docs/components/alert">Alert</a>
-  </li>
-  <li>
     <a href="/docs/components/animated-image">Animated Image</a>
   </li>
   <li>
@@ -70,6 +67,9 @@
   </li>
   <li>
     <a href="/docs/components/button-group">Button Group</a>
+  </li>
+  <li>
+    <a href="/docs/components/callout">Callout</a>
   </li>
   <li>
     <a href="/docs/components/card">Card</a>
