+++ docs/docs/components/callout.md
@@ -1,0 +1,147 @@
+---
+title: Callout
+description: Callouts are used to display important messages inline.
+layout: component.njk
+---
+
+```html {.example}
+<wa-callout>
+  <wa-icon slot="icon" name="circle-info" variant="regular"></wa-icon>
+  This is a standard callout. You can customize its content and even the icon.
+</wa-callout>
+```
+
+{% raw %}
+```jsx {.react}
+import WaCallout from '@shoelace-style/shoelace/dist/react/callout';
+import WaIcon from '@shoelace-style/shoelace/dist/react/icon';
+
+const App = () => (
+  <WaCallout>
+    <WaIcon slot="icon" name="circle-info" variant="regular" />
+    This is a standard callout. You can customize its content and even the icon.
+  </WaCallout>
+);
+```
+{% endraw %}
+
+## Examples
+
+### Variants
+
+Set the `variant` attribute to change the callout's variant.
+
+```html {.example}
+<wa-callout variant="brand">
+  <wa-icon slot="icon" name="circle-info" variant="regular"></wa-icon>
+  <strong>This is super informative</strong><br />
+  You can tell by how pretty the callout is.
+</wa-callout>
+
+<br />
+
+<wa-callout variant="success">
+  <wa-icon slot="icon" name="circle-check" variant="regular"></wa-icon>
+  <strong>Your changes have been saved</strong><br />
+  You can safely exit the app now.
+</wa-callout>
+
+<br />
+
+<wa-callout variant="neutral">
+  <wa-icon slot="icon" name="gear" variant="regular"></wa-icon>
+  <strong>Your settings have been updated</strong><br />
+  Settings will take effect on next login.
+</wa-callout>
+
+<br />
+
+<wa-callout variant="warning">
+  <wa-icon slot="icon" name="triangle-exclamation" variant="regular"></wa-icon>
+  <strong>Your session has ended</strong><br />
+  Please login again to continue.
+</wa-callout>
+
+<br />
+
+<wa-callout variant="danger">
+  <wa-icon slot="icon" name="circle-exclamation" variant="regular"></wa-icon>
+  <strong>Your account has been deleted</strong><br />
+  We're very sorry to see you go!
+</wa-callout>
+```
+
+{% raw %}
+```jsx {.react}
+import WaCallout from '@shoelace-style/shoelace/dist/react/callout';
+import WaIcon from '@shoelace-style/shoelace/dist/react/icon';
+
+const App = () => (
+  <>
+    <WaCallout variant="brand">
+      <WaIcon slot="icon" name="circle-info" variant="regular" />
+      <strong>This is super informative</strong>
+      <br />
+      You can tell by how pretty the callout is.
+    </WaCallout>
+
+    <br />
+
+    <WaCallout variant="success">
+      <WaIcon slot="icon" name="circle-check" variant="regular" />
+      <strong>Your changes have been saved</strong>
+      <br />
+      You can safely exit the app now.
+    </WaCallout>
+
+    <br />
+
+    <WaCallout variant="neutral">
+      <WaIcon slot="icon" name="gear" variant="regular" />
+      <strong>Your settings have been updated</strong>
+      <br />
+      Settings will take effect on next login.
+    </WaCallout>
+
+    <br />
+
+    <WaCallout variant="warning">
+      <WaIcon slot="icon" name="triangle-exclamation" variant="regular" />
+      <strong>Your session has ended</strong>
+      <br />
+      Please login again to continue.
+    </WaCallout>
+
+    <br />
+
+    <WaCallout variant="danger">
+      <WaIcon slot="icon" name="circle-exclamation" variant="regular" />
+      <strong>Your account has been deleted</strong>
+      <br />
+      We're very sorry to see you go!
+    </WaCallout>
+  </>
+);
+```
+{% endraw %}
+
+### Without Icons
+
+Icons are optional. Simply omit the `icon` slot if you don't want them.
+
+```html {.example}
+<wa-callout variant="brand"> Nothing fancy here, just a simple callout. </wa-callout>
+```
+
+{% raw %}
+```jsx {.react}
+import WaCallout from '@shoelace-style/shoelace/dist/react/callout';
+
+const App = () => (
+  <WaCallout variant="brand">
+    Nothing fancy here, just a simple callout.
+  </WaCallout>
+);
+```
+{% endraw %}
+
