--- src/components/alert/alert.test.ts
@@ -1,308 +1,0 @@
-import { aTimeout, expect, fixture, html, oneEvent } from '@open-wc/testing';
-import { clickOnElement, moveMouseOnElement } from '../../internal/test.js';
-import { queryByTestId } from '../../internal/test/data-testid-helpers.js';
-import { resetMouse } from '@web/test-runner-commands';
-import sinon from 'sinon';
-import type WaAlert from './alert.js';
-import type WaIconButton from '../icon-button/icon-button.js';
-
-const getAlertContainer = (alert: WaAlert): HTMLElement => {
-  return alert.shadowRoot!.querySelector<HTMLElement>('[part="base"]')!;
-};
-
-const expectAlertToBeVisible = (alert: WaAlert): void => {
-  const alertContainer = getAlertContainer(alert);
-  const style = window.getComputedStyle(alertContainer);
-  expect(style.display).not.to.equal('none');
-  expect(style.visibility).not.to.equal('hidden');
-  expect(style.visibility).not.to.equal('collapse');
-};
-
-const expectAlertToBeInvisible = (alert: WaAlert): void => {
-  const alertContainer = getAlertContainer(alert);
-  const style = window.getComputedStyle(alertContainer);
-  expect(style.display, 'alert should be invisible').to.equal('none');
-};
-
-const expectHideAndAfterHideToBeEmittedInCorrectOrder = async (alert: WaAlert, action: () => void | Promise<void>) => {
-  const hidePromise = oneEvent(alert, 'wa-hide');
-  const afterHidePromise = oneEvent(alert, 'wa-after-hide');
-  let afterHideHappened = false;
-  oneEvent(alert, 'wa-after-hide').then(() => (afterHideHappened = true));
-
-  action();
-
-  await hidePromise;
-  expect(afterHideHappened).to.be.false;
-
-  await afterHidePromise;
-  expectAlertToBeInvisible(alert);
-};
-
-const expectShowAndAfterShowToBeEmittedInCorrectOrder = async (alert: WaAlert, action: () => void | Promise<void>) => {
-  const showPromise = oneEvent(alert, 'wa-show');
-  const afterShowPromise = oneEvent(alert, 'wa-after-show');
-  let afterShowHappened = false;
-  oneEvent(alert, 'wa-after-show').then(() => (afterShowHappened = true));
-
-  action();
-
-  await showPromise;
-  expect(afterShowHappened).to.be.false;
-
-  await afterShowPromise;
-  expectAlertToBeVisible(alert);
-};
-
-const getCloseButton = (alert: WaAlert): WaIconButton | null | undefined =>
-  alert.shadowRoot?.querySelector<WaIconButton>('[part="close-button"]');
-
-describe('<wa-alert>', () => {
-  let clock: sinon.SinonFakeTimers | null = null;
-
-  afterEach(async () => {
-    clock?.restore();
-    // eslint-disable-next-line
-    await resetMouse().catch(() => {});
-  });
-
-  it('renders', async () => {
-    const alert = await fixture<WaAlert>(html`<wa-alert open>I am an alert</wa-alert>`);
-
-    expectAlertToBeVisible(alert);
-  });
-
-  it('is accessible', async () => {
-    const alert = await fixture<WaAlert>(html`<wa-alert open>I am an alert</wa-alert>`);
-
-    await expect(alert).to.be.accessible();
-  });
-
-  describe('alert visibility', () => {
-    it('should be visible with the open attribute', async () => {
-      const alert = await fixture<WaAlert>(html`<wa-alert open>I am an alert</wa-alert>`);
-
-      expectAlertToBeVisible(alert);
-    });
-
-    it('should not be visible without the open attribute', async () => {
-      const alert = await fixture<WaAlert>(html` <wa-alert>I am an alert</wa-alert>`);
-
-      expectAlertToBeInvisible(alert);
-    });
-
-    it('should emit wa-show and wa-after-show when calling show()', async () => {
-      const alert = await fixture<WaAlert>(html` <wa-alert>I am an alert</wa-alert>`);
-
-      expectAlertToBeInvisible(alert);
-
-      await expectShowAndAfterShowToBeEmittedInCorrectOrder(alert, async () => await alert.show());
-    });
-
-    it('should emit wa-hide and wa-after-hide when calling hide()', async () => {
-      const alert = await fixture<WaAlert>(html` <wa-alert open>I am an alert</wa-alert>`);
-
-      await expectHideAndAfterHideToBeEmittedInCorrectOrder(alert, async () => await alert.hide());
-    });
-
-    it('should emit wa-show and wa-after-show when setting open = true', async () => {
-      const alert = await fixture<WaAlert>(html` <wa-alert>I am an alert</wa-alert> `);
-
-      await expectShowAndAfterShowToBeEmittedInCorrectOrder(alert, async () => {
-        alert.open = true;
-        await alert.updateComplete;
-      });
-    });
-
-    it('should emit wa-hide and wa-after-hide when setting open = false', async () => {
-      const alert = await fixture<WaAlert>(html` <wa-alert open>I am an alert</wa-alert> `);
-
-      await expectHideAndAfterHideToBeEmittedInCorrectOrder(alert, async () => {
-        alert.open = false;
-        await alert.updateComplete;
-      });
-    });
-  });
-
-  describe('close button', () => {
-    it('shows a close button if the alert has the closable attribute', () => async () => {
-      const alert = await fixture<WaAlert>(html` <wa-alert open closable>I am an alert</wa-alert> `);
-      const closeButton = getCloseButton(alert);
-
-      expect(closeButton).to.be.visible;
-    });
-
-    it('clicking the close button closes the alert', () => async () => {
-      const alert = await fixture<WaAlert>(html` <wa-alert open closable>I am an alert</wa-alert> `);
-      const closeButton = getCloseButton(alert);
-
-      await expectHideAndAfterHideToBeEmittedInCorrectOrder(alert, async () => {
-        await clickOnElement(closeButton!);
-      });
-    });
-  });
-
-  describe('toast', () => {
-    const getToastStack = (): HTMLDivElement | null => document.querySelector<HTMLDivElement>('.wa-toast-stack');
-
-    const closeRemainingAlerts = async (): Promise<void> => {
-      const toastStack = getToastStack();
-      if (toastStack?.children) {
-        for (const element of toastStack.children) {
-          await (element as WaAlert).hide();
-        }
-      }
-    };
-
-    beforeEach(async () => {
-      await closeRemainingAlerts();
-    });
-
-    it('can be rendered as a toast', async () => {
-      const alert = await fixture<WaAlert>(html`<wa-alert>I am an alert</wa-alert>`);
-
-      await expectShowAndAfterShowToBeEmittedInCorrectOrder(alert, async () => await alert.toast());
-      const toastStack = getToastStack();
-      expect(toastStack).to.be.visible;
-      expect(toastStack?.firstChild).to.be.equal(alert);
-    });
-
-    it('resolves only after being closed', async () => {
-      const alert = await fixture<WaAlert>(html`<wa-alert closable>I am an alert</wa-alert>`);
-
-      const afterShowEvent = oneEvent(alert, 'wa-after-show');
-      let toastPromiseResolved = false;
-      alert.toast().then(() => (toastPromiseResolved = true));
-
-      await afterShowEvent;
-      expect(toastPromiseResolved).to.be.false;
-
-      const closePromise = oneEvent(alert, 'wa-after-hide');
-      const closeButton = getCloseButton(alert);
-      clickOnElement(closeButton!);
-
-      await closePromise;
-      await aTimeout(0);
-
-      expect(toastPromiseResolved).to.be.true;
-    });
-
-    const expectToastStack = () => {
-      const toastStack = getToastStack();
-      expect(toastStack).not.to.be.null;
-    };
-
-    const expectNoToastStack = () => {
-      const toastStack = getToastStack();
-      expect(toastStack).to.be.null;
-    };
-
-    const openToast = async (alert: WaAlert): Promise<void> => {
-      const openPromise = oneEvent(alert, 'wa-after-show');
-      alert.toast();
-      await openPromise;
-    };
-
-    const closeToast = async (alert: WaAlert): Promise<void> => {
-      const closePromise = oneEvent(alert, 'wa-after-hide');
-      const closeButton = getCloseButton(alert);
-      await clickOnElement(closeButton!);
-      await closePromise;
-      await aTimeout(0);
-    };
-
-    it('deletes the toast stack after the last alert is done', async () => {
-      const container = await fixture<HTMLElement>(
-        html`<div>
-          <wa-alert data-testid="alert1" closable>alert 1</wa-alert>
-          <wa-alert data-testid="alert2" closable>alert 2</wa-alert>
-        </div>`
-      );
-
-      const alert1 = queryByTestId<WaAlert>(container, 'alert1');
-      const alert2 = queryByTestId<WaAlert>(container, 'alert2');
-
-      await openToast(alert1!);
-
-      expectToastStack();
-
-      await openToast(alert2!);
-
-      expectToastStack();
-
-      await closeToast(alert1!);
-
-      expectToastStack();
-
-      await closeToast(alert2!);
-
-      expectNoToastStack();
-    });
-  });
-
-  describe('timer controlled closing', () => {
-    it('closes after a predefined amount of time', async () => {
-      clock = sinon.useFakeTimers();
-      const alert = await fixture<WaAlert>(html` <wa-alert open duration="3000">I am an alert</wa-alert>`);
-
-      expectAlertToBeVisible(alert);
-
-      clock.tick(2999);
-
-      expectAlertToBeVisible(alert);
-
-      await expectHideAndAfterHideToBeEmittedInCorrectOrder(alert, () => {
-        clock?.tick(1);
-      });
-    });
-
-    it('resets the closing timer after mouse-over', async () => {
-      clock = sinon.useFakeTimers();
-      const alert = await fixture<WaAlert>(html` <wa-alert open duration="3000">I am an alert</wa-alert>`);
-
-      expectAlertToBeVisible(alert);
-
-      clock.tick(1000);
-
-      await moveMouseOnElement(alert);
-
-      clock.tick(2999);
-
-      expectAlertToBeVisible(alert);
-
-      await expectHideAndAfterHideToBeEmittedInCorrectOrder(alert, () => {
-        clock?.tick(1);
-      });
-    });
-
-    it('resets the closing timer after opening', async () => {
-      clock = sinon.useFakeTimers();
-      const alert = await fixture<WaAlert>(html` <wa-alert duration="3000">I am an alert</wa-alert>`);
-
-      expectAlertToBeInvisible(alert);
-
-      clock.tick(1000);
-
-      const afterShowPromise = oneEvent(alert, 'wa-after-show');
-      alert.show();
-      await afterShowPromise;
-
-      clock.tick(2999);
-
-      await expectHideAndAfterHideToBeEmittedInCorrectOrder(alert, () => {
-        clock?.tick(1);
-      });
-    });
-  });
-});
-
-it('Should properly render alert variants', async () => {
-  const variants = ['brand', 'success', 'neutral', 'warning', 'danger'];
-
-  for (const variant of variants) {
-    const alert = await fixture<WaAlert>(html`<wa-alert variant="${variant}" open>I am an alert</wa-alert>`);
-
-    const alertContainer = getAlertContainer(alert);
-    expect(alertContainer).to.have.class(`alert--${variant}`);
-  }
-});
